+++ packages/frontend/pages/settings/components/sidebar/SettingsSidebar.tsx
@@ -1,12 +1,15 @@
 import NextLink from 'next/link';
 import { Box, BoxProps, Button, VStack } from '@chakra-ui/react';
 import { IoReturnUpBack } from '@react-icons/all-files/io5/IoReturnUpBack';
+import { useIsTouchscreen } from '@utils';
 import { SettingsCategory } from './SettingsCategory';
 import { SettingsLink } from './SettingsLink';
 
 export type SettingsSidebarProps = BoxProps;
 
 export const SettingsSidebar: React.FC<SettingsSidebarProps> = (props) => {
+  const isTouchscreen = useIsTouchscreen();
+
   return (
     <VStack w="full" align="flex-start" spacing={3} {...props}>
       <NextLink href="/" passHref>
@@ -23,9 +26,11 @@ export const SettingsSidebar: React.FC<SettingsSidebarProps> = (props) => {
         <SettingsLink href="/settings/appearance">Appearance</SettingsLink>
       </SettingsCategory>
 
-      <SettingsCategory title="Collections">
-        <SettingsLink href="/settings/organize">Organize</SettingsLink>
-      </SettingsCategory>
+      {!isTouchscreen && (
+        <SettingsCategory title="Collections">
+          <SettingsLink href="/settings/organize">Organize</SettingsLink>
+        </SettingsCategory>
+      )}
 
       <SettingsCategory title="Account">
         <SettingsLink href="/settings/info">Info</SettingsLink>