+++ packages/frontend/theme/fonts.ts
@@ -1,66 +1,66 @@
 export const fontFaces = `
-/* nunito-regular - latin */
+/* nunito-regular - latin-ext_latin */
 @font-face {
   font-family: 'Nunito';
   font-style: normal;
   font-weight: 400;
   src: local(''),
-       url('/fonts/nunito-v20-latin-regular.woff2') format('woff2'),
-       url('/fonts/nunito-v20-latin-regular.woff') format('woff');
+       url('/fonts/nunito-v25-latin-ext_latin-regular.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/nunito-v25-latin-ext_latin-regular.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 
-/* nunito-700 - latin */
+/* nunito-700 - latin-ext_latin */
 @font-face {
   font-family: 'Nunito';
   font-style: normal;
   font-weight: 700;
   src: local(''),
-       url('/fonts/nunito-v20-latin-700.woff2') format('woff2'),
-       url('/fonts/nunito-v20-latin-700.woff') format('woff');
+       url('/fonts/nunito-v25-latin-ext_latin-700.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/nunito-v25-latin-ext_latin-700.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 
-/* ubuntu-regular - latin */
+/* ubuntu-regular - latin-ext_latin */
 @font-face {
   font-family: 'Ubuntu';
   font-style: normal;
   font-weight: 400;
   font-display: block;
   src: local(''),
-       url('/fonts/ubuntu-v20-latin-regular.woff2') format('woff2'),
-       url('/fonts/ubuntu-v20-latin-regular.woff') format('woff');
+       url('/fonts/ubuntu-v20-latin-ext_latin-regular.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/ubuntu-v20-latin-ext_latin-regular.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 
-/* ubuntu-700 - latin */
+/* ubuntu-700 - latin-ext_latin */
 @font-face {
   font-family: 'Ubuntu';
   font-style: normal;
   font-weight: 700;
   font-display: block;
   src: local(''),
-       url('/fonts/ubuntu-v20-latin-700.woff2') format('woff2'),
-       url('/fonts/ubuntu-v20-latin-700.woff') format('woff');
+       url('/fonts/ubuntu-v20-latin-ext_latin-700.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/ubuntu-v20-latin-ext_latin-700.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 
-/* lato-regular - latin */
+/* lato-regular - latin-ext_latin */
 @font-face {
   font-family: 'Lato';
   font-style: normal;
   font-weight: 400;
   font-display: block;
   src: local(''),
-       url('/fonts/lato-v23-latin-regular.woff2') format('woff2'),
-       url('/fonts/lato-v23-latin-regular.woff') format('woff');
+       url('/fonts/lato-v23-latin-ext_latin-regular.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/lato-v23-latin-ext_latin-regular.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 
-/* lato-700 - latin */
+/* lato-700 - latin-ext_latin */
 @font-face {
   font-family: 'Lato';
   font-style: normal;
   font-weight: 700;
   font-display: block;
   src: local(''),
-       url('/fonts/lato-v23-latin-700.woff2') format('woff2'),
-       url('/fonts/lato-v23-latin-700.woff') format('woff');
+       url('/fonts/lato-v23-latin-ext_latin-700.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/lato-v23-latin-ext_latin-700.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 
 /* OpenDyslexic Regular */
@@ -113,68 +113,68 @@ export const fontFaces = `
        url('/fonts/OpenDyslexicMono-Regular.otf') format('opentype');
 }
 
-/* source-code-pro-regular - latin */
+/* source-code-pro-regular - latin-ext_latin */
 @font-face {
   font-family: 'Source Code Pro';
   font-style: normal;
   font-weight: 400;
   src: local(''),
-       url('/fonts/source-code-pro-v21-latin-regular.woff2') format('woff2'),
-       url('/fonts/source-code-pro-v21-latin-regular.woff') format('woff');
+       url('/fonts/source-code-pro-v21-latin-ext_latin-regular.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/source-code-pro-v21-latin-ext_latin-regular.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 
-/* source-code-pro-700 - latin */
+/* source-code-pro-700 - latin-ext_latin */
 @font-face {
   font-family: 'Source Code Pro';
   font-style: normal;
   font-weight: 700;
   src: local(''),
-       url('/fonts/source-code-pro-v21-latin-700.woff2') format('woff2'),
-       url('/fonts/source-code-pro-v21-latin-700.woff') format('woff');
+       url('/fonts/source-code-pro-v21-latin-ext_latin-700.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/source-code-pro-v21-latin-ext_latin-700.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 
-/* fira-mono-700 - latin */
-@font-face {
-  font-family: 'Fira Mono';
-  font-style: normal;
-  font-weight: 700;
-  font-display: block;
-  src: local(''),
-       url('/fonts/fira-mono-v14-latin-700.woff2') format('woff2'),
-       url('/fonts/fira-mono-v14-latin-700.woff') format('woff');
-}
-
-/* fira-mono-regular - latin */
+/* fira-mono-regular - latin-ext_latin */
 @font-face {
   font-family: 'Fira Mono';
   font-style: normal;
   font-weight: 400;
   font-display: block;
   src: local(''),
-       url('/fonts/fira-mono-v14-latin-regular.woff2') format('woff2'),
-       url('/fonts/fira-mono-v14-latin-regular.woff') format('woff');
+       url('/fonts/fira-mono-v14-latin-ext_latin-regular.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/fira-mono-v14-latin-ext_latin-regular.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 
-/* ubuntu-mono-regular - latin */
+/* fira-mono-700 - latin-ext_latin */
+@font-face {
+  font-family: 'Fira Mono';
+  font-style: normal;
+  font-weight: 700;
+  font-display: block;
+  src: local(''),
+       url('/fonts/fira-mono-v14-latin-ext_latin-700.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/fira-mono-v14-latin-ext_latin-700.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
+}
+
+/* ubuntu-mono-regular - latin-ext_latin */
 @font-face {
   font-family: 'Ubuntu Mono';
   font-style: normal;
   font-weight: 400;
   font-display: block;
   src: local(''),
-       url('/fonts/ubuntu-mono-v15-latin-regular.woff2') format('woff2'),
-       url('/fonts/ubuntu-mono-v15-latin-regular.woff') format('woff');
+       url('/fonts/ubuntu-mono-v15-latin-ext_latin-regular.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/ubuntu-mono-v15-latin-ext_latin-regular.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 
-/* ubuntu-mono-700 - latin */
+/* ubuntu-mono-700 - latin-ext_latin */
 @font-face {
   font-family: 'Ubuntu Mono';
   font-style: normal;
   font-weight: 700;
   font-display: block;
   src: local(''),
-       url('/fonts/ubuntu-mono-v15-latin-700.woff2') format('woff2'),
-       url('/fonts/ubuntu-mono-v15-latin-700.woff') format('woff');
+       url('/fonts/ubuntu-mono-v15-latin-ext_latin-700.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
+       url('/fonts/ubuntu-mono-v15-latin-ext_latin-700.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
 }
 `;